+++ docs/changelog.md
@@ -5,3 +5,7 @@ All notable changes to this project will be documented below.
 #### geospatial.read_geotif
 
 * v0.1dev: spectral = **geospatial.read_geotif**(*filename, bands="B,G,R"*)
+
+#### geospatial.transform_points
+
+* v0.1dev: coord = **geospatial.transform_points**(*img, geojson*)